--- docs/src/components/data/types.ts
@@ -1,22 +1,0 @@
-const sourceTypes = [
-    'vector',
-    'raster',
-    'raster-dem',
-    'geojson',
-    'image',
-    'video'
-];
-
-const layerTypes = [
-    'background',
-    'fill',
-    'line',
-    'symbol',
-    'raster',
-    'circle',
-    'fill-extrusion',
-    'heatmap',
-    'hillshade'
-];
-
-export {sourceTypes, layerTypes};
