+++ docs/src/components/property.tsx
