+++ docs/src/components/sdk-support-table/sdk-support-table.module.scss
