+++ docs/src/routes/layers.tsx
@@ -1,9 +1,9 @@
-import {SolidMd} from '~/utils/SolidMd';
+import {SolidMd} from '~/components/SolidMd';
 import ref from '../reference/latest';
 import {Items} from '../components/style-spec/specs/items';
 import {AppropriateImage} from '~/components/style-spec/appropriate-image';
-import Caption from '~/components/style-spec/caption';
-import combineItems from '~/components/style-spec/combine-items';
+import Caption from '~/components/caption';
+import combineItems from '~/util/combine-items';
 
 function Layers() {
